I tried

+++ src/main/java/com/rentalsystem/swp/Repositories/UserRepository.java
@@ -1,13 +1,13 @@
 package com.rentalsystem.swp.Repositories;
 
-import com.rentalsystem.swp.dao.UserProfile;
+import com.rentalsystem.swp.models.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserProfile, Integer> {
+public interface UserRepository extends JpaRepository<UserProfile, Long> {
     UserProfile getById(long id);
     Optional<UserProfile> findById(long id);
     Optional<UserProfile> findByEmail(String email);
